fix: correct type mismatches for ToolStateRunning.Input, OAuth.Expires, and SessionForkParams.MessageID
ToolStateRunning.Input is spec'd as any JSON value (not just objects),
OAuth.Expires is spec'd as number (not integer), and SessionForkParams.MessageID
is optional per spec (no required array in request body).

+++ config.go
@@ -1522,7 +1522,7 @@ type OAuth struct {
 	Type    AuthType `json:"type"`
 	Refresh string   `json:"refresh"`
 	Access  string   `json:"access"`
-	Expires int64    `json:"expires"`
+	Expires float64  `json:"expires"`
 }
 
 func (OAuth) implementsAuthSetParamsAuthUnion() {}
